requirements_gemini_test.txt
# Requirements for Gemini Image Editing Test
google-generativeai>=0.8.0
Pillow>=10.4.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: For development
python-dotenv>=1.0.0
